Fixed input from URL so that items are put into the list in order, no matter when the AJAX requests are sent out. This is being done by assigning each input request an index in a hash table. All requests for an index are put into a list. Once all AJAX requests are complete, the hash table is transformed into a flat list, and all items are added to the playlist.
The AJAX manager is still being used to ensure that all AJAX requests are completed before unblocking the UI, as well as to ensure that playlist requests (e.g. SoundCloud playlists) put the songs in the playlist into the correct order. If this functionality can be written in a different way, the AJAX manager will no longer be necessary.