Bug fix time!
Fixed window hash not changing if tracks are removed, shuffled, or moved.

Fixed shuffling sometimes causing the current track to switch to a track with a different ID during shuffle.

Fixed issue where tracks could be moved in list but the list would not be altered (bug in jQuery UI, sortable module) if the user moved an item to the end of the list by moving their mouse below the y-boundary of #tracks. The fix looks less elegant if the user's screen is not full of tracks (moved track floats in #main wherever user's mouse is), but they shouldn't notice anything otherwise.