Radical UI change! The entire UI is now entirely on the right-hand side. I realized this was possible when I looked at the layout and saw two things:
1. A lot of the top header space was going unused.
2. A lot of the right side space was going unused.

Since I couldn't move the video to the header without resizing the video container or the header (neither of which would be acceptable), I decided to just move everything over to the right side.

I'm happy with it, more or less. I may need to fiddle with spacing and the size of the media source bar and seek bar, but that can be adjusted later. For now, I am pleased that there is a lot more real estate left on the screen for what matters--the playlist itself. I am tempted to move the right side to the left side so the controls are closer to the delete button, but again, that is not important right now.

Also, the sliders are now using the jQuery UI slider library. The library I wrote was buggy, and I would rather go with a solution verified by someone else. Plus, I managed to retain the same look and feel of the sliders, another win.