Router no longer relies on the jQuery batch AJAX plug-in for loading a playlist. Jesus, I was dumb to think I needed queues or batches or anything other than knowing how many AJAX requests would need to be run. Maybe I'm being too hard on myself. This line of thinking relies on the assumption that loading the media from the playlist requires only one AJAX request per item. Then again, it's not the case that _variable_ requests are needed per item, so I guess it was kinda silly. I might turn my jQuery batch plugin into a jQuery request counter (not sure if that's useful).
On the other hand, I am not yet sure if I need to do any special real-time handling of adding playlist items. The fundamental assumption is that users can't add media fast enough for the media to be added out of order. However, I may need to challenge that assumption.

Overall, working with real-time functionality is tough, and I'm glad I'm taking the time to work on this project. I've made some dumb mistakes, but as I work with JavaScript I realize that I am learning a great deal.