Major overhaul, part 1. In an effort to make the application more fit to handle asynchronous events, I am causing the router to work on deferred objects. This makes the router more complicated, but it also makes reactions to asynchronous events more predictable.
For instance, if a user adds a YouTube link and a SoundCloud link with little delay between the two actions, which gets executed first? Before there was no guarantee the YouTube video would go in the playlist first. Now, the action to add each track goes into a queue. When an action is executed, the action at the head of the queue is dequeued and executed, and the router checks to see if there are any more actions that need to be executed.

The problem is that the Reddit action does not order the way in which it adds links to the playlist, so that will have to be reworked. For now I'm holding off, because it's late at night. Blegh.