Fixed major bug in the router where adding a playlist item would return a Deferred object that resolved before the item was actually added to the playlist. Unfortunately, I had to fix this issue by adding another layer of Deferreds, this time to resolve after the playlist actions are executed. I am now in the awkward position of relying on Deferreds for a lot of event chaining. I fear that changing the playlist to not use these Deferreds chains would be complicated, but at least the thing works. Here is the current chain of Deferreds.
1. Deferred A is created in router.addResource. This Deferred will resolve when the AJAX call(s) to gather the web media are complete.
2. Deferred B is created in router.addResource. This Deferred will resolve when the resource adding action is complete.
3. Deferred A resolves if the requested resource can be captured, and otherwise Deferred A and Deferred B reject.
4. If Deferred A resolved, continue. Deferred A is added to router.actionQueue. If Deferred A is the first Deferred in the queue, an event is chained to it that tells the router to empty the queue by executing the functions stored in the actionQueue Deferreds in the resolve method. The emptying is done in router.executeActionQueueItems.
5. In router.executeActionQueueItems, the action and postAction functions are executed associated with each Deferred's resolve data. This resolves Deferred B.

I feel like this is insane. No sane human being would want to write code like this, but I am unable to see another way. Maybe I need to take a break from this part of the project? Everything seems to work, and now the Deferred returned by router.addResource is guaranteed to only resolve after something has been completely routed. One part of me says that this is the only way to get that to work, but another says that there must be a simpler way. I feel like I have implemented some perverse version of asynchronous method queue chaining. I read and reread Dustin Diaz's blog article on the subject and I am practically convinced I did everything the wrong way. Deferreds are not a panacea, and I tried to use them as such.

http://www.dustindiaz.com/async-method-queues/

And yet, his code doesn't guarantee any order. It only allows asynchronous functionality to be chained. Perhaps I went about the playlist the wrong way fundamentally. Right now I attempt to guarantee order in added playlist items, which are added all at once. Another approach would be to insert items as soon as they arrive and insert them into the playlist when the media is available. Users expect the former because that's the way offline media is loaded, but the latter is easier to implement.

I am not happy, but the code works. I can ship this (well, upload it to my server), and may God have mercy on my soul.