On a whim, I decided to add a new feature... the ability to route any URL that contains YouTube or SoundCloud embedded players.
There is a couple potential problems here with this feature. It uses a lot of bandwidth, because I have to download an entire page through the server (because JS restricts cross-browser requests), route that to the client, and then parse the output data.

This could be mitigated by using the YQL server, although I am loath to use a 3rd party solution because it gives me less control. However, my use of YQL would only become a problem if A) Yahoo shuts down--a real possibility due to their financial troubles--or B) I get so large that I have to change the way I use YQL. Possibility B is less likely IMO, so I might as well utilize YQL for as long as possible.