Intermediate, experimental commit that creates a search form and reorganizes the page.
This commit moves the playlist to the right side of the page under the video, gets rid of the "Add Media Source" button, and adds a search form. The user is not yet able to add search results to the playlist, but this feature is forthcoming. Also, the main layout will need to undergo a radical change very soon to accommodate the search results.