Added search results formatting and the ability to add tracks from search results.
Also, results can now be added to the playlist, albeit through a crude system where the playlist makes another request to YouYube/SoundCloud. It works, but the next step is to make that step reliant on the search results data. Also, after that tracks should be immediately playable from search results.