Revamped UI is now largely functional.
I am now dropping the experimental/unstable label, because the new UI is mostly complete. There are still a few issues, but I am extremely happy with the way the new UI looks. There are still some sizing/color issues that need to be sorted out, but overall the result is impressive.