Corrected YouTube issue and link column issue in table.
For some videos, YouTube would complain if the video palyer was < 200px. I bumped the height to that amount to avoid that, although there is now a border around videos. Such is compromise. I also added slightly more space to the link column in the track tables so it would not be so squished.