Fix for problems moving tracks.
There was an issue where tracks moved to the end of the playlist would be interpreted as outside the table, causing issues from detecting the new index of the track. Additionally, there was a typo in the code to move non-playing tracks, causing the current track to be reset to the first track.