Added database support and caching search results.
Database support is currently used to add data, but data retrieval is not yet implemented. It will be used to quickly verify media IDs when a user wants to create a new playlist ID.

Caching is used to save time and bandwidth that would ordinarily be used to search YouTube or SoundCloud for media. The cache is very short-lived, ensuring that users will not see old data and that the server will not be overwhelmed.