Adding DB related features for playlists.
So far playlist strings can be searched in the DB and playlist IDs can be found. Additionally, playlists can be saved. Work needs to be done on the client side to get a new playlist ID when the playlist changes, as well as on the server side to facilitate these requests/responses. Caching would work extremely well here given enough traffic.