Changed playlist to be responsive to the user pressing the back and forward buttons.
To change the playlist without refreshing, I am using the facilities of history.js. When the playlist is changing the state of the page, it marks a variable stating it is busy. Every time the state is changed, this variable is checked. If not marked, the playlist is reset and changed to whatever is fetched from the API.