Fixed problem where gaps would appear in the autoincrement ID.
This was due to my ignorance of "INSERT IGNORE" updating the autoincrement upon any insert, even ignored ones. Because several playlist actions made it possible to call the API several times for the same playlist, gaps would appear when the same track would be submitted multiple times and only saved once. This means the problem had nothing to do with MySQL connection pooling.