Revamp the TrackPlaylist class. The playlist was previously having problems where the current track would be set before the track views were actually created. The most elegant solution is for the track views to be added first, by causing the playlist view to listen for an earlier event to add the track views. Is is accomplished by listening to events in the "tracks" namespace.
Additionally, the isEmpty method has been removed in favor of isLoaded, which is a more useful method. Most uses of isEmpty are used to say, "Do this if the playlist is not empty." It is simpler to just ask if the playlist is loaded, which also implies that the playlist has set its currentMedia property.