Make progress on model event triggering.
Playlist now does not trigger play or resume events on its own, which will be replaced soon. Also, since currently the playlist triggers progress events when a track is playing, this functionality will be offloaded to the models. Additionally, a bug was fixed for YouTube-callbacks where the event arguments would include the arguments of the callback, rather than the arguments passed to the model method.