Add more volume tests and fix various volume bugs.
Some issues were present where volume changes or muting would occur but checking the status of the change would not register the change. This is due to the Playlist storing some status at the wrong time, which is probably not a good long-term strategy (e.g. storing a muted property vs checking if the current track is muted.)