Add tests related to adding playlist tracks, and fix a bug in the search results cache.
The search results cache was storing results in a way the server did not expect. When retrieving results from the cache, the server would serve a non-existent property from the returned object, _tracks_. This would propagate as null, causing the search results collection to choke on any subsequent searches for a previously searched keyword.

The acceptance test script has also been moved into a closure to prevent any leakage to the global environment that might affect the tests.