A cornucopia of changes related to database functionality.
An install target is now specified for make, with a corresponding setup script in the bin directory. This script is used to setup the database Muxamp needs to store playlist and media data.

Additionally, database pools can be created with a specific name and/or options. This was introduced because of the need to specify that the setup script should be allowed to execute multiple statements, something normally not needed by Muxamp.

Some documentation has also been updated.