TrackPlaylist#currentMedia is now a function.
This is because of a condition (potentially a bug) where the playlist identifies a track as the current track, but this track is removed from the playlist. When the playlist executes currentMedia.play(), a track outside the playlist starts playing. To prevent this from happening, currentMedia is a function that returns the track at index TrackPlaylist#currentTrack. This behavior is simpler than the previous implementation, where both the current media and the track number were recorded. Now, only the current track number is recorded.