Fix YouTube tracks not triggering the "end" event, and visual responsiveness of the timebar.
YouTube tracks would refuse to autoplay the next track on play end, because the TrackPlaylist class waits for the "end" event before going on to the next track.

The visual adjustment made to the timebar does not actually make Muxamp work faster to load the next one, but it LOOKS faster because the animation happens while nothing would otherwise be happening.