Default track number for TrackPlaylist is now -1.
The previous default was 0, but since this number is also a possible value for #currentTrack after TrackPlaylist is initialized, a situation could arise where the playlist would be loaded, but the 'track' event would never be triggered.

Specifically, this would happen every time an existing playlist was loaded. Since the PlaylistView depends on this track to decorate the current playlist track, the first track in a loaded playlist would never be highlighted--only tracks played after it, since the 'track' event is triggered whenever a new track becomes the current track for the playlist.