Improve performance of playlist verification query in MySQL.
Using an IN condition in SQL is more efficient than using a chain of OR statements, since the IN condition is an easily optimizable operation checking for membership in a set, while OR causes the query runner to execute potentially arbitrary expressions (if they cannot be optimized).

A quick experiment shows that, for equivalent queries verifying playlist tracks (pop size 100, N=1), the IN version requires 90 microseconds for initialization, 17 microsecs for optimization, 17 microsecs for statistics, and 19 microsecs for preparation. The OR version, by contract, needs 144, 140, 349, and 109 microseconds for initialization, optimization, statistics, and preparation respectively. In this one instance, the optimizable parts of the query can be made 5x faster.

This will surely make Muxamp better able to scale by taking less time to run otherwise expensive queries without loss of correctness.