Fix a bug where switching from a playing track a different track plays the wrong track.
This was caused by an event problem related to YouTube tracks. Muxamp relies YouTube tracks to emit the "end" event before switching to the next track. Because stopping a track causes this event to be emitted, TrackPlaylist#setCurrentTrack attempted to stop listening to the "end" event. The intention was to override the automatic transition from track X to X + 1 when the user selects track Y.

Unfortunately, the event was not properly ignored, causing the X + 1 track to play. The test environment did not catch it because it was misconfigured; the timeout was set to 10 seconds, but waiting for all the required events to register before completing the test required 14.