Add a way for Muxamp to recover current track state after going back and forth in the browser history.
Using track index alone to store playlist state is not a comprehensive solution, because it is a pointer, not a "primary key" in the sense of uniqueness.

For instance, imagine the following sequence of events:

1. Current playlist track index is 3 of 4.
2. Track 2 is removed. (I: 3 -> 2)
3. User hits back in browser. (I: 2)
4. User his forward in browser. (I: 2)

The current track should be back to 3, but it is instead 2. Because we have not tied the current playing track to its index, we have no way of knowing which track is the current track without checking each one.