Add Jamendo support.
I chose Jamendo because I have been a fan of them for a while, and their API looked like it was easy to integrate.

The process of adding Jamendo was educational to me, as it showed me the difficulty of the process of adding media sources to Muxamp. Although Muxamp does a good job of playing media, the integration process leaves something to be desired. Code had to be added in several different places to make everything work. Muxamp needs to be more modular--adding a media source should be a server-side operation, such that the only client side code needed would be the model needed to play that media.